+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add Arps decline rate to natural geologic hydrogen model
 - Added linearized hydrogen fuel cell model
 - Added load following optimization dispatch
 - Added simple dispatch calculations to `StorageAutoSizingModel`